+++ lib/main_mainapp.dart
@@ -1,3 +1,4 @@
+import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
@@ -13,6 +14,9 @@ import 'package:stopandgo/routes/app_routes.dart';
 import 'core/config/flavor_config.dart';
 import 'core/network/api_client.dart';
 
+late FirebaseAnalytics firebaseAnalytics;
+late FirebaseAnalyticsObserver firebaseObserver;
+
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
@@ -24,6 +28,21 @@ Future<void> main() async {
     bundleId: 'app.stopandgomx.main',
     organizationId: 1,
   );
+
+  firebaseAnalytics = FirebaseAnalytics.instance;
+  firebaseObserver = FirebaseAnalyticsObserver(analytics: firebaseAnalytics);
+
+  await firebaseAnalytics.setUserProperty(
+    name: 'flavor',
+    value: FlavorConfig.I.flavor.name,
+  );
+
+  if (FlavorConfig.I.organizationId != null) {
+    await firebaseAnalytics.setUserProperty(
+      name: 'organization_id',
+      value: FlavorConfig.I.organizationId.toString(),
+    );
+  }
 
   FirebaseMessaging.onBackgroundMessage(firebaseMessagingBackgroundHandler);
   await NotificationService.initialize();